feat: added lexical analysis on lexer.c.
added the lexical analysis implementation to the lexer.c file. The lexical analysis is responsible for identifying the tokens present in the source code. Functions have been added to create tokens, display lexical errors, add characters to a word, check if a word is reserved and perform the lexical analysis itself. Explanatory comments have also been added to make the code easier to understand.